Karta med lite knappar
Just nu äre så att användarn får dra markören till den stad den tror är
rätt. När användarn klicka på gissaknappen postas markörens koordinater
till controllern. I den metoden kan vi ju sen lägga in att vi kollar om
svaret är rätt då! :) Men utseendet på allt kan vi ju ändra på sen om vi
hinner! :+1:

+++ Kartastrof/Kartastrof/Models/Tbl_Capital.cs
@@ -14,8 +14,8 @@ namespace Kartastrof.Models
 
         public string Ca_Name { get; set; }
 
-        public int? Ca_Longitude { get; set; }
+        public float? Ca_Longitude { get; set; }
 
-        public int? Ca_Latitude { get; set; }
+        public float? Ca_Latitude { get; set; }
     }
 }
